@@ -1,16 +1,20 @@
 <?php
 
-namespace Youngyezi\Captcha\Facades;
+namespace LumenUtils\Captcha\Facades;
 
 
 use Illuminate\Support\Facades\Facade;
+
 /**
- * @see \Mews\Captcha
+ * @see \LumenUtils\Captcha
  */
 class Captcha extends Facade
 {
     /**
      * @return string
      */
-    protected static function getFacadeAccessor() { return 'captcha'; }
+    protected static function getFacadeAccessor()
+    {
+        return 'captcha';
+    }
 }
